It's possible again to select assays
