+++ src/components/pages/HomePage.vue
@@ -1,37 +1,36 @@
 <template>
-    <v-container fluid>
-        <v-row>
-            <v-col>
-                <!-- <p>
-                    Welcome to Unipept Desktop! Please select a valid project folder or create a new project before continuing.
-                </p> -->
-                <recent-projects></recent-projects>
-            </v-col>
-            <v-divider vertical></v-divider>
-            <v-col>
-                <div style="display: flex; flex-direction: column; align-items: center;">
-                    <img src="@/assets/logo.svg" style="max-width: 200px;"/>
-                    <span class="logo-headline">Unipept Desktop</span>
-                    <span class="logo-subline">Version 0.0.1</span>
+    <div class="homepage-container">
+        <v-container fluid>
+            <v-row>
+                <v-col>
+                    <recent-projects></recent-projects>
+                </v-col>
+                <v-divider vertical></v-divider>
+                <v-col>
+                    <div style="display: flex; flex-direction: column; align-items: center;">
+                        <img src="@/assets/logo.svg" style="max-width: 200px;"/>
+                        <span class="logo-headline">Unipept Desktop</span>
+                        <span class="logo-subline">Version 0.0.1</span>
 
-                    <div class="project-actions">
-                        <div @click="createProject()">
-                            <v-icon>mdi-folder-plus-outline</v-icon>
-                            <span>Create new project</span>
-                        </div>
-                        <div>
-                            <v-icon>mdi-folder-download-outline</v-icon>
-                            <span>Import project</span>
+                        <div class="project-actions">
+                            <div @click="createProject()">
+                                <v-icon>mdi-folder-plus-outline</v-icon>
+                                <span>Create new project</span>
+                            </div>
+                            <div>
+                                <v-icon>mdi-folder-download-outline</v-icon>
+                                <span>Import project</span>
+                            </div>
                         </div>
                     </div>
-                </div>
-            </v-col>
-        </v-row>
-        <v-snackbar v-model="errorSnackbarVisible" bottom :timeout="0" color="error">
-            {{ errorMessage }}
-            <v-btn dark text @click="errorSnackbarVisible = false">Close</v-btn>
-        </v-snackbar>
-    </v-container>
+                </v-col>
+            </v-row>
+            <v-snackbar v-model="errorSnackbarVisible" bottom :timeout="0" color="error">
+                {{ errorMessage }}
+                <v-btn dark text @click="errorSnackbarVisible = false">Close</v-btn>
+            </v-snackbar>
+        </v-container>
+    </div>
 </template>
 
 <script lang="ts">
@@ -60,6 +59,8 @@ export default class HomePage extends Vue {
             try {
                 const projectMng: ProjectManager = new ProjectManager();
                 const project: Project = await projectMng.initializeProject(chosenPath[0]);
+                this.$store.dispatch("setProject", project);
+                this.$router.push("/analysis/single");
             } catch (err) {
                 if (err instanceof IOException) {
                     this.showError(
@@ -109,5 +110,9 @@ export default class HomePage extends Vue {
     .project-actions .v-icon {
         position: relative;
         top: -2px;
+    }
+
+    .homepage-container, .homepage-container .container {
+        display: flex;
     }
 </style>
